Use Long for a timeout value in milliseconds

+++ cli/src/main/kotlin/com/malinskiy/marathon/cli/args/FileConfiguration.kt
@@ -31,7 +31,7 @@ data class FileConfiguration(
         var includeSerialRegexes: Collection<Regex>?,
         var excludeSerialRegexes: Collection<Regex>?,
 
-        var testOutputTimeoutMillis: Int?,
+        var testOutputTimeoutMillis: Long?,
         var debug: Boolean?,
 
         var vendorConfiguration: FileVendorConfiguration?
